requirements.txt
boto3==1.14.14
line-bot-sdk==1.16.0